scx_wd40: [non-functional-changes] refactor domain load balancing logic into separate compilation units
