+++ scheds/rust/scx_wd40/build.rs
@@ -8,6 +8,8 @@ fn main() {
         .unwrap()
         .enable_intf("src/bpf/intf.h", "bpf_intf.rs")
         .enable_skel("src/bpf/main.bpf.c", "bpf")
+        .add_source("src/bpf/lb_domain.bpf.c")
+        .add_source("src/bpf/common.bpf.c")
         .add_source("../../../lib/sdt_task.bpf.c")
         .add_source("../../../lib/sdt_alloc.bpf.c")
         .compile_link_gen()
